move "selectable email listing" to manual backup constructor.
